Implement internal bridge to return room’s usernames
