+++ src/definition/messages/IMessage.ts
@@ -1,4 +1,4 @@
-import { IRoom, Room } from '../rooms';
+import { IRoom } from '../rooms';
 import { IUser } from '../users';
 import { IMessageAttachment } from './IMessageAttachment';
 import { IMessageReactions } from './IMessageReaction';
@@ -19,31 +19,4 @@ export interface IMessage {
     reactions?: IMessageReactions;
     groupable?: boolean;
     customFields?: { [key: string]: any };
-}
-
-export class Message implements IMessage {
-    public id?: string;
-    public sender: IUser;
-    public text?: string;
-    public createdAt?: Date;
-    public updatedAt?: Date;
-    public editor?: IUser;
-    public editedAt?: Date;
-    public emoji?: string;
-    public avatarUrl?: string;
-    public alias?: string;
-    public attachments?: Array<IMessageAttachment>;
-    public reactions?: IMessageReactions;
-    public groupable?: boolean;
-    public customFields?: { [key: string]: any };
-    private _ROOM: Room;
-    public get room(): Room {
-        return this._ROOM;
-    }
-    public set room(room) {
-        this._ROOM = new Room(room);
-    }
-    public constructor(message: IMessage) {
-        Object.assign(this, message);
-    }
 }
