+++ src/definition/rooms/index.ts
@@ -4,12 +4,11 @@ import { IPreRoomCreateExtend } from './IPreRoomCreateExtend';
 import { IPreRoomCreateModify } from './IPreRoomCreateModify';
 import { IPreRoomCreatePrevent } from './IPreRoomCreatePrevent';
 import { IPreRoomDeletePrevent } from './IPreRoomDeletePrevent';
-import { IRoom, Room } from './IRoom';
+import { IRoom } from './IRoom';
 import { RoomType } from './RoomType';
 
 export {
     IRoom,
-    Room,
     RoomType,
     IPostRoomCreate,
     IPostRoomDeleted,
