+++ src/server/bridges/IInternalBridge.ts
@@ -1,0 +1,3 @@
+export interface IInternalBridge {
+    getUsernamesOfRoomById(roomId: string): Array<string>;
+}
